Made some improvement on index file (#273)
Links with target blank were missing rel attribute with noreferrer value and the referred information was passed to the website being linked. Also, the page can be controlled from the new website being linked because no opener value was missing from the rel attribute. This is a security approach because links to the cross-origin destinations are unsafe. Read more on below link:

https://developers.google.com/web/tools/lighthouse/audits/noopener 

Fixed some a tag missing > character and image tag missing alt attribute.